@@ -1,8 +1,7 @@
-module.exports = {
+export default {
   clearMocks: true,
   moduleFileExtensions: ['js', 'ts'],
   modulePaths: ['<rootDir>'],
-  setupFiles: ['<rootDir>/jest.setup.js'],
   testEnvironment: 'node',
   testMatch: ['**/*.test.ts'],
   testTimeout: 1000,
