@@ -4,9 +4,8 @@ export interface Element<T> {
 }
 
 // When there isn't time to implement a real heap in an interview, and the @datastructures-js/priority-queue library
-// isn't available, you can use binary search insert in place of a real heap.  This will be O(1) for popping, but it
-// will be O(n) for insert (even though the binary search is O(logn), we do have to resize the array at O(n)).
-export class BinarySearchMaxHeap<T> {
+// isn't available, just insert and sort.  This will be O(nlogn) on insert but it's quick and dirty.
+export class SimpleMaxHeap<T> {
   private readonly heap: Element<T>[];
 
   private readonly compare: (a: number, b: number) => number;
@@ -16,43 +15,30 @@ export class BinarySearchMaxHeap<T> {
     this.compare = compare ?? ((a, b) => a - b);
   }
 
-  public enqueue(value: T, priority?: number): void {
+  public enqueue(value: T, priority?: number): SimpleMaxHeap<T> {
     const element: Element<T> = {
       value,
       priority: this.getPriority(value, priority)
     };
-
-    let left = 0;
-    let right = this.heap.length;
-    while (left < right) {
-      const mid = Math.floor((left + right) / 2);
-      const a = this.heap[mid];
-      const b = element;
-      // You can use < here for non-stable inserts, if you don't care.  Switch to <= for stable inserts.
-      if (this.compare(a.priority, b.priority) < 0) {
-        left = mid + 1;
-      } else {
-        right = mid;
-      }
-    }
-
-    this.heap.splice(left, 0, element);
+    this.heap.push(element);
+    this.heap.sort((a, b) => this.compare(a.priority, b.priority));
+    return this;
   }
 
-  public dequeue(): Element<T> {
+  public dequeue(): T {
     if (this.isEmpty()) {
       throw new Error('heap is empty');
     }
 
-    return this.heap.pop()!;
+    return this.heap.pop()!.value;
   }
 
-  public front(): Element<T> {
+  public front(): T {
     if (this.isEmpty()) {
       throw new Error('heap is empty');
     }
 
-    return this.heap[this.heap.length - 1];
+    return this.heap[this.heap.length - 1].value;
   }
 
   public size(): number {
